requirements.txt
alembic==1.16.2
anyio==4.9.0
certifi==2025.6.15
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
sniffio==1.3.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
fastapi==0.115.14
pydantic==2.11.1
pydantic-settings==2.2.1
pydantic_core==2.33.0
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
datasets
scikit-learn
transformers